style: extract types and path logic from zon utils
