+++ src/utils/zon/zon.ts
@@ -1,120 +1,19 @@
-import { rainbow, Cubehelix } from "./color.ts";
-import { CodeStats, Languages } from "./tokei.ts";
-
-export const LineType = {
-  blanks: "blanks",
-  code: "code",
-  comments: "comments",
-} as const;
-export type LineType = keyof typeof LineType;
-
-export const NODE_TYPE = {
-  FILE: "FILE",
-  FOLDER: "FOLDER",
-  GROUP: Symbol("GROUP"),
-} as const;
-export type NODE_TYPE = keyof typeof NODE_TYPE;
-
-export type SegmentName = string | typeof NODE_TYPE.GROUP;
-export type Path = SegmentName[];
-
-type Colors = {
-  base: string;
-  highlighted: string;
-  pressed: string;
-};
-
-type NodeBase = {
-  stats: CodeStats;
-  path: Path;
-  name: SegmentName;
-  numberOfLines: number;
-  firstLine: number;
-  colors: Colors;
-  depth: number;
-};
-
-export type File = NodeBase & {
-  type: typeof NODE_TYPE.FILE;
-  height: 0;
-};
-
-export type Folder = NodeBase & {
-  type: typeof NODE_TYPE.FOLDER;
-  children: Node[];
-  height: number;
-};
-
-export type Group = NodeBase & {
-  type: typeof NODE_TYPE.GROUP;
-  groupedChildren: Node[];
-  height: number;
-};
-
-export type Node = File | Folder | Group;
-
-export function getProjectName(projectPath: string): string {
-  const projectPathSegments = projectPath.split("/");
-
-  return projectPathSegments[projectPathSegments.length - 1];
-}
-
-/**
- * For user-facing purposes
- * @param name
- * @returns
- */
-export function getDisplayName(name: SegmentName): string {
-  if (name === NODE_TYPE.GROUP) {
-    // TODO: i18n
-    return "Smaller items";
-  }
-
-  if (typeof name === "symbol") {
-    throw new Error(
-      `Can't convert segment name "${name.toString()}" into a string, unknown symbol type`,
-    );
-  }
-
-  return name;
-}
-
-/**
- * Use only for internal & debugging purposes
- * @param path
- * @returns
- */
-export function getPathString(path: Path | null): string {
-  if (path == null) {
-    return "null";
-  }
-
-  return path
-    .map((segment) =>
-      segment === NODE_TYPE.GROUP ? 'Symbol("group")' : segment,
-    )
-    .join("/");
-}
-
-export function arePathsEqual(left: Path | null, right: Path | null): boolean {
-  if (left == null && right == null) {
-    return true;
-  }
-
-  if (left == null || right == null) {
-    return false;
-  }
-
-  if (left.length !== right.length) {
-    return false;
-  }
-
-  return left.every((segment, i) => segment === right[i]);
-}
-
-export function getParentPath(path: Path) {
-  return path.slice(0, -1);
-}
+import { rainbow, Cubehelix } from "../color.ts";
+import { CodeStats, Languages } from "../tokei.ts";
+import {
+  type Colors,
+  type LineType,
+  type Node,
+  NODE_TYPE,
+  type Path,
+  type SegmentName,
+} from "./types.ts";
+import {
+  getProjectName,
+  getPathString,
+  getParentPath,
+  arePathsEqual,
+} from "./path.ts";
 
 export function createTree(
   projectPath: string,
